cleaned out typeIndex unused code
